% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.source.R
\name{as.source}
\alias{as.source}
\title{Return miniMeta analysis as source code}
\usage{
as.source(x)
}
\arguments{
\item{x}{An object of class \code{miniMeta}}
}
\value{
A character vector of length one, containing R code that 
loads the data, runs the meta-analysis, and plots a forest plot. 
You can save this in a text file using \code{\link[base]{writeLines}}.
}
\description{
Returns an entire miniMeta analysis in an R source code format. 
This provides a basis for further processing the results exported 
from miniMeta, using R code, in order to perform more elaborate or
more specific analyses.
}
