% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{example_analyses}
\alias{example_analyses}
\alias{example_miniMeta_obs}
\alias{example_miniMeta_rct}
\title{Example miniMeta analyses}
\format{
Objects of class \code{miniMeta}

An object of class \code{miniMeta} (inherits from \code{list}) of length 4.

An object of class \code{miniMeta} (inherits from \code{list}) of length 4.
}
\usage{
example_miniMeta_obs

example_miniMeta_rct
}
\description{
These are example miniMeta meta-analyses, with study data taken from
\href{https://pubmed.ncbi.nlm.nih.gov/24587664/}{Lytras et al, 2014}.
Object \code{example_miniMeta_rct} contains a meta-analysis of Randomized 
Controlled Trials (RCTs), and \code{example_miniMeta_obs} a meta-analysis 
of observational studies.
}
\references{
Lytras T, Nikolopoulos G, Bonovas S. World J Gastroenterol 
2014;20(7):1858-70 (\href{https://pubmed.ncbi.nlm.nih.gov/24587664/}{PubMed})
}
\keyword{datasets}
