% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.miniMeta.R
\name{forest.miniMeta}
\alias{forest.miniMeta}
\title{Forest plot for miniMeta objects}
\usage{
\method{forest}{miniMeta}(x)
}
\arguments{
\item{x}{An object of class \code{miniMeta}}
}
\description{
Draws a forest plot for a miniMeta object using the options 
stored in the object
}
