% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.miniMeta.R
\name{is.miniMeta}
\alias{is.miniMeta}
\title{Is this a miniMeta object?}
\usage{
is.miniMeta(x)
}
\arguments{
\item{x}{An object of class \code{miniMeta}}
}
\value{
\code{TRUE} if it is a valid miniMeta object, \code{FALSE} if it is not.
}
\description{
This function checks whether this is a valid miniMeta object
}
\examples{
is.miniMeta(example_miniMeta_obs) # returns TRUE
is.miniMeta(example_miniMeta_rct) # returns TRUE

}
