% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseArguments.R
\name{parseArguments}
\alias{parseArguments}
\title{Parse arguments from a comma-separated list}
\usage{
parseArguments(x)
}
\arguments{
\item{x}{A character vector (of length one) containing the arguments.
All should be named.}
}
\value{
A named list of arguments, or an object of class "try-error" on failure.
}
\description{
Read a comma-separated list of arguments (as a character string), 
parse them, and return as a named R list. This function is used in
miniMeta to parse arguments for forest.meta() given as a string.
}
