% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOptions.R
\name{plotOptions}
\alias{plotOptions}
\title{Get forest plot options from miniMeta object}
\usage{
plotOptions(x)
}
\arguments{
\item{x}{An object of class \code{miniMeta}}
}
\value{
A named list of arguments corresponding to the arguments of 
\code{\link[meta]{forest.meta}}.
}
\description{
This function returns the forest plot options stored in a miniMeta 
object, as a named list of arguments, for further processing.
}
\examples{
\donttest{
# m should be a miniMeta object
fopts <- plotOptions(m)
do.call(forest, c(x=m, fopts))
}

}
